# priv/static/swagger.json
{
  "definitions": {
    "Category": {
      "description": "A category of the application",
      "properties": {
        "code": {
          "description": "Category code",
          "type": "string"
        },
        "description": {
          "description": "Category description",
          "type": "text"
        },
        "id": {
          "description": "Category ID",
          "type": "string"
        },
        "inserted_at": {
          "description": "When was the activity initially inserted",
          "format": "datetime",
          "type": "string"
        },
        "name": {
          "description": "Category name",
          "type": "string"
        },
        "updated_at": {
          "description": "When was the activity last updated",
          "format": "datetime",
          "type": "string"
        }
      },
      "required": [
        "code",
        "description",
        "name"
      ],
      "title": "Category",
      "type": "object"
    },
    "CreateCategoryRequest": {
      "description": "POST body for creating a category",
      "example": {
        "code": "AR900",
        "description": "ISO 9000 is a set of quality control and quality management standards established by the International Organization for Standardization. They can be applied in any type of organization or activity oriented to the production of goods or services.",
        "name": "ISO 9000"
      },
      "properties": {
        "category": {
          "$ref": "#/definitions/Category",
          "description": "The category details"
        }
      },
      "title": "CreateCategoryRequest",
      "type": "object"
    },
    "CreateCategoryResponse": {
      "description": "Response schema for single category",
      "example": {
        "data": {
          "code": "AR900",
          "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
          "id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
          "name": "Operation",
          "words": []
        }
      },
      "properties": {
        "data": {
          "$ref": "#/definitions/Category",
          "description": "The category details created"
        }
      },
      "title": "CreateCategoryResponse",
      "type": "object"
    },
    "CreateCategoryResponseErrors": {
      "description": "Response errors from create category",
      "example": {
        "code": "can't be blank",
        "description": "can't be blank",
        "name": "can't be blank"
      },
      "title": "CreateCategoryResponseErrors",
      "type": "object"
    },
...
    "Error": {
      "description": "Error responses from the API",
      "properties": {
        "error": {
          "description": "The message of the error raised",
          "type": "string"
        }
      },
      "required": [
        "error"
      ],
      "title": "Errors",
      "type": "object"
    },
    "ShowCategoriesEmptyResponse": {
      "description": "Response error categories empty",
      "example": {
        "error": "There are no categories"
      },
      "title": "ShowCategoriesEmptyResponse",
      "type": "object"
    },
    "ShowCategoriesResponse": {
      "description": "Response schema for multiple categories created",
      "example": {
        "categories": [
          {
            "code": "AR900",
            "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
            "id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
            "name": "Operation",
            "words": [
              {
                "category_id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
                "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
                "id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
                "image": {
                  "file_name": "operation.png",
                  "updated_at": "2022-12-16T15:55:28"
                },
                "name": "Operation"
              },
              {
                "category_id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
                "description": "For the ISO-9000 standard, quality is the degree to which a set of inherent characteristics meets the requirements, being: Quality characteristic: inherent characteristic of a product, process or system related to a requirement.",
                "id": "df48a064-75d5-4be7-a773-7003169277e6",
                "image": {
                  "file_name": "quality.png",
                  "updated_at": "2022-12-16T15:55:28"
                },
                "name": "Quality"
              }
            ]
          },
          {
            "category_id": "5a10f256-51cd-4e91-844f-13d17a65ad57",
            "description": "For the ISO-9000 standard, quality is the degree to which a set of inherent characteristics meets the requirements, being: Quality characteristic: inherent characteristic of a product, process or system related to a requirement.",
            "id": "df48a064-75d5-4be7-a773-7003169277e6",
            "image": {
              "file_name": "quality.png",
              "updated_at": "2022-12-16T15:55:28"
            },
            "name": "Quality"
          }
        ],
        "first": 1,
        "has_next": false,
        "has_prev": false,
        "last": 2,
        "next_page": 2,
        "page": 1,
        "prev_page": 0,
        "total_count": 2
      },
      "properties": {
        "categories": {
          "description": "The categories details",
          "items": {
            "$ref": "#/definitions/Category"
          },
          "type": "array"
        }
      },
      "title": "ShowCategoriesResponse",
      "type": "object"
    },
    "ShowCategoryResponse": {
      "description": "Response schema of a single category",
      "example": {
        "data": {
          "category_id": "5a10f256-51cd-4e91-844f-13d17a65ad57",
          "description": "For the ISO-9000 standard, quality is the degree to which a set of inherent characteristics meets the requirements, being: Quality characteristic: inherent characteristic of a product, process or system related to a requirement.",
          "id": "df48a064-75d5-4be7-a773-7003169277e6",
          "image": {
            "file_name": "quality.png",
            "updated_at": "2022-12-16T15:55:28"
          },
          "name": "Quality"
        }
      },
      "properties": {
        "data": {
          "$ref": "#/definitions/Category",
          "description": "The category found"
        }
      },
      "title": "ShowCategoryResponse",
      "type": "object"
    },
...
    "UpdateCategoryRequest": {
      "description": "Request params to update a category",
      "example": {
        "name": "ISO 26000"
      },
      "properties": {
        "data": {
          "description": "The category params",
          "items": {
            "$ref": "#/definitions/Word"
          },
          "type": "array"
        }
      },
      "title": "UpdateCategoryRequest",
      "type": "object"
    },
    "UpdateCategoryResponse": {
      "description": "Response schema of a single category",
      "example": {
        "data": {
          "code": "AR900",
          "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
          "id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
          "name": "Operation",
          "words": [
            {
              "category_id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
              "description": "For the ISO-9000 standard, quality is the degree to which a set of inherent characteristics meets the requirements, being: Quality characteristic: inherent characteristic of a product, process or system related to a requirement.",
              "id": "df48a064-75d5-4be7-a773-7003169277e6",
              "image": {
                "file_name": "quality.png",
                "updated_at": "2022-12-16T15:55:28"
              },
              "name": "Quality"
            }
          ]
        }
      },
      "properties": {
        "data": {
          "description": "The category updated",
          "items": {
            "$ref": "#/definitions/Word"
          },
          "type": "array"
        }
      },
      "title": "UpdateCategoryResponse",
      "type": "object"
    },
...
    "Word": {
      "description": "A word of the application",
      "properties": {
        "category_id": {
          "description": "ID of category to which the word belongs",
          "type": "string"
        },
        "description": {
          "description": "Word description",
          "type": "text"
        },
        "id": {
          "description": "Word ID",
          "type": "string"
        },
        "image": {
          "description": "Word image",
          "type": "string"
        },
        "inserted_at": {
          "description": "When was the activity initially inserted",
          "format": "datetime",
          "type": "string"
        },
        "name": {
          "description": "Word name",
          "type": "string"
        },
        "updated_at": {
          "description": "When was the activity last updated",
          "format": "datetime",
          "type": "string"
        }
      },
      "required": [
        "category_id",
        "description",
        "name"
      ],
      "title": "Word",
      "type": "object"
    }
  },
  "host": "localhost:4000",
  "info": {
    "contact": {
      "email": "orlando@cordage.io",
      "name": "Orlando Gomez"
    },
    "description": "API Documentation for GreenHouse API v1",
    "termsOfService": "Open for public",
    "title": "Words GreenHouse API",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/categories": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "deprecated": false,
        "description": "Create a new category in the database",
        "operationId": "GlossaryWeb.CategoryController.create",
        "parameters": [
          {
            "description": "The category data",
            "in": "body",
            "name": "word",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCategoryRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreateCategoryResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/CreateCategoryResponseErrors"
            }
          }
        },
        "summary": "Add a new category",
        "tags": [
          "Category"
        ]
      }
    },
...
    "/api/v1/categories/{id}": {
      "delete": {
        "description": "Delete a category by ID",
        "operationId": "GlossaryWeb.CategoryController.delete",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content - Deleted Successfully"
          }
        },
        "summary": "Delete a specific category",
        "tags": [
          "Category"
        ]
      },
      "get": {
        "deprecated": false,
        "description": "Return JSON with an especific category",
        "operationId": "GlossaryWeb.CategoryController.show",
        "parameters": [
          {
            "description": "Category ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Category created OK",
            "schema": {
              "$ref": "#/definitions/ShowCategoryResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Show a specific category",
        "tags": [
          "Category"
        ]
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "description": "Update all attributes of a category",
        "operationId": "GlossaryWeb.CategoryController.update",
        "parameters": [
          {
            "description": "Category ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "The category details to update",
            "in": "body",
            "name": "category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCategoryRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "205": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/UpdateCategoryResponse"
            }
          },
          "400": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Update a category",
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/categories?page={page}&page_size={size}": {
      "get": {
        "deprecated": false,
        "description": "Return JSON with all categories listed in the database",
        "operationId": "GlossaryWeb.CategoryController.index",
        "parameters": [
          {
            "description": "Page number to be displayed",
            "in": "path",
            "name": "page",
            "required": true,
            "type": "integer",
            "x-example": 2
          },
          {
            "description": "Total number of entries per page",
            "in": "path",
            "name": "size",
            "required": true,
            "type": "integer",
            "x-example": 10
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ShowCategoriesResponse"
            }
          },
          "204": {
            "description": "No Categories",
            "schema": {
              "$ref": "#/definitions/ShowCategoriesEmptyResponse"
            }
          }
        },
        "summary": "All Categories",
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/search": {
      "get": {
        "deprecated": false,
        "description": "Returns JSON with all categories matching listed in the database",
        "operationId": "GlossaryWeb.CategoryController.search_category",
        "parameters": [
          {
            "description": "Page number to be displayed",
            "in": "path",
            "name": "page",
            "required": true,
            "type": "integer",
            "x-example": 2
          },
          {
            "description": "Total number of entries per page",
            "in": "path",
            "name": "size",
            "required": true,
            "type": "integer",
            "x-example": 10
          },
          {
            "description": "The field you want to find",
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ShowCategoriesResponse"
            }
          },
          "204": {
            "description": "No users",
            "schema": {
              "$ref": "#/definitions/ShowCategoriesEmptyResponse"
            }
          }
        },
        "summary": "All categories in the search",
        "tags": [
          "Category"
        ]
      }
    }
  },
  "swagger": "2.0"
}